Fix alignment issues with list on homepage
When padding was added to ingredients on recipe pages, the ul type
selector was used. This adjusted the list on the homepage because it is
also a ul type and its position was altered using position: relative
rather than padding. So the padding was added and then its position was
adjusted relative to it's new position. A class was assigned to the
ingredients lists and the class selector was used instead to return the
list on the homepage to its original position.