Resize border around ingredients
The border was going across the entire webpage instead of being limited
to just surrounding the content. So I changed the height and with of the
div element to fit-content and set the left margin to 300px to align
with the rest of the page. I then adjusted the padding of the
ingredients and ingredients header to match up to the new left margin.